backend/requirements.txt
fastapi
uvicorn
pydantic[dotenv]
pydantic-settings
pyjwt
passlib
python-multipart
python-dotenv
cryptography>=42.0.0
sqlalchemy
python-jose[cryptography]
httpx
pyyaml
